Attempting to fix windows path issues
