+++ crates/nu_plugin_polars/src/dataframe/command/core/save/arrow.rs
@@ -6,7 +6,7 @@ use polars::prelude::{IpcWriter, SerWriter};
 use polars_io::ipc::IpcWriterOptions;
 
 use crate::{
-    command::core::Resource,
+    command::core::resource::Resource,
     values::{NuDataFrame, NuLazyFrame},
 };
