+++ crates/nu_plugin_polars/src/dataframe/command/core/save/parquet.rs
@@ -7,7 +7,7 @@ use polars::prelude::ParquetWriter;
 use polars_io::parquet::write::ParquetWriteOptions;
 
 use crate::{
-    command::core::Resource,
+    command::core::resource::Resource,
     values::{NuDataFrame, NuLazyFrame},
 };
